refactor(retention): split data-retention page + round presets up
Two follow-ups from PR #4147 review (tracked in #4506):

1. Round retention presets UP relative to the human label so the
   selection covers the full period plus a small buffer instead of
   underselling by 1-10 days. "1 year" was 364d, now 371d; "2 years"
   728->735, "5 years" 1820->1827.

2. Split the 1185-line src/pages/settings/data-retention.tsx into the
   page shell + per-concern modules under src/components/data-retention/:
   - constants.ts        category/scope labels, presets, units, sentinels
   - format.ts           formatBytes, formatDays
   - grouping.ts         RetentionRuleRow/Group + group/render helpers
   - AddOverrideDrawer.tsx
   - RetroactiveProgressCard.tsx
   - ApplyToExistingConfirmDialog.tsx
   - RetentionAndUsageCard.tsx

   Page is now 575 lines (-51%) holding only the shell + data
   orchestration. No behavior change; data-retention unit tests still
   green (8 files, 66 tests).
