+++ langwatch/src/components/data-retention/format.ts
@@ -1,0 +1,11 @@
+export function formatBytes(bytes: number): string {
+  if (bytes === 0) return "0 B";
+  const units = ["B", "KB", "MB", "GB", "TB"];
+  const i = Math.floor(Math.log(bytes) / Math.log(1024));
+  const value = bytes / Math.pow(1024, i);
+  return `${value.toFixed(2)} ${units[i]}`;
+}
+
+export function formatDays(days: number): string {
+  return days === 0 ? "Indefinite" : `${days} days`;
+}
